app/requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
flask-sqlalchemy==2.5.1
psycopg2==2.9.3
Flask-Cors==3.0.10
bcrypt==3.2.0
Flask-JWT-Extended==4.3.1
Flask-Login==0.4.1
Flask-User==1.0.2.2
email-validator==1.1.3
sqlalchemy-filters==0.12.0
sqlalchemy-filters-plus==1.0.4
gunicorn==20.1.0
pandas==1.4.0
requests==2.27.1
googlemaps==4.6.0
geopy==2.2.0
usps-api==0.5